# apps/api/openapi.yaml
openapi: 3.1.0
info:
  title: NOZOMI 用车平台 API
  version: 0.1.0
  description: |
    接口定义先于实现：新增或修改接口时先改本文件，再写代码。
    `apps/api/src/openapi.test.ts` 会校验「已注册的路由」和本文件的 `paths` 完全一致。

    约定：
    - 出错时统一返回 `{ error: { code, message, details } }`（见 `ErrorResponse`）。
      唯一的例外是 `/health` 的 503：为了让监控直接看到是哪一项出了问题，仍返回 `HealthResponse`。
    - 所有接口共有的错误（不在每个接口下重复列出）：
      - 400 `BAD_REQUEST`：请求体不是合法 JSON、URL 里有非法的百分号编码、发来的根本不是 HTTP 请求
      - 400 `VALIDATION_FAILED`：参数校验未通过，`details.issues` 列出每一项
      - 404 `NOT_FOUND`：路径或请求方法不存在
      - 408 `REQUEST_TIMEOUT`：请求迟迟没有发完
      - 413 `PAYLOAD_TOO_LARGE`：请求体超过 1MB
      - 415 `UNSUPPORTED_MEDIA_TYPE`：请求内容类型不支持
      - 431 `REQUEST_HEADER_FIELDS_TOO_LARGE`：请求头超过 16KB
      - 500 `INTERNAL_ERROR`：未预期的服务端错误，`message` 不含内部细节

      其中 408、431 和「不是 HTTP 请求」的 400 在解析阶段就被拒绝，应答之后连接会被服务端关闭。
    - 金额用最小货币单位整数，百分比用基点整数。
    - 租户接口的 `tenant_id` 只从登录令牌读取；请求参数和请求体里的 `tenant_id` 一律被忽略。
    - `/platform/v1` 和 `/tenant/v1` 的字段名用 snake_case，时间用带时区的 ISO 8601。
      （`/health` 是 M0-05 定下的运维接口，字段名保持原样。）
    - 登录：平台员工用 `/platform/v1/auth/login`，租户用户用 `/tenant/v1/auth/login`，
      拿到访问令牌后放在请求头 `Authorization: Bearer <令牌>`。令牌 8 小时过期，没有刷新令牌，过期后重新登录。
      两类令牌互不通用：平台令牌访问租户接口、租户令牌访问平台接口，都是 401。
      停用账号、退出登录、凭重置令牌改了密码之后，已经发出的令牌立即失效。
      租户被暂停不影响它的用户登录（暂停的租户还要继续履约已有订单），`tenant.status` 会告诉前端当前状态。
    - 所有字符串输入都不能包含 NUL 字符（`\u0000`、查询串里的 `%00`），也不能包含孤立的代理字符（被截断的半个表情符号），
      否则 400 `VALIDATION_FAILED`。
...
      什么都没有改动，原样重试即可。主数据的写接口会先自动重试，极少会真的返回它。
...
servers:
...
paths:
...
  /platform/v1/master/cities:
    get:
      tags: [主数据]
      operationId: listCities
      summary: 城市列表
      description: |
        需要操作 `master_data.read`（每个平台角色都有）。默认按创建时间从早到晚（创建时间相同的按编号，顺序稳定）；`sort=code` 按编码。
        `total` 是符合筛选条件的总数；`next_cursor` 不为 null 表示还有下一页。`total` 和这一页的内容读自同一时刻的数据。
        不带 `status` 时返回全部（启用的和停用的）。
      security:
        - platformToken: []
      parameters:
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/MasterStatusFilter"
        - $ref: "#/components/parameters/MasterCodeFilter"
        - $ref: "#/components/parameters/MasterKeyword"
        - $ref: "#/components/parameters/MasterSort"
        - $ref: "#/components/parameters/UpdatedSince"
        - $ref: "#/components/parameters/CountryCodeFilter"
      responses:
        "200":
          description: 一页城市
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [items, next_cursor, total]
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/City"
                  next_cursor:
                    type: [string, "null"]
                  total:
                    type: integer
                    minimum: 0
                    description: 符合筛选条件的总数（和翻到第几页无关）
    post:
      tags: [主数据]
      operationId: createCity
      summary: 新增城市
      description: |
        需要操作 `master_data.manage`（主数据运营、超级管理员）。
        编码格式 `CTY-国家码-序号`（如 `CTY-JP-TYO`），其中的国家码必须等于 `country_code`。
        时区必须是 IANA 时区名（如 `Asia/Tokyo`；不接受缩写、固定偏移、UTC）。坐标是 WGS84，保留小数点后 6 位。
        新增的城市直接是启用的。
        写审计日志（`resource` = `city`，`action` = `create`，`after` 是完整内容）。
      security:
        - platformToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CityCreate"
      responses:
        "201":
          description: 新增的城市
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/City"
        "409":
          description: |
            - `CODE_TAKEN`：这个编码已被使用
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /platform/v1/master/cities/{id}:
    get:
      tags: [主数据]
      operationId: getCity
      summary: 查看一个城市
      description: 需要操作 `master_data.read`。
      security:
        - platformToken: []
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: 城市
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/City"
        "404":
          $ref: "#/components/responses/ResourceNotFound"
    patch:
      tags: [主数据]
      operationId: updateCity
      summary: 修改城市
      description: |
        需要操作 `master_data.manage`。只改请求体里带了的字段。
        必须用请求头 `If-Match` 带上当前的 `version`：和库里的不一致说明别人先改过了，返回 409 `VERSION_CONFLICT`，不会互相覆盖。
        `code`、`country_code` 创建后不能改。`name`、`center`、`boundary` 带了就是整体替换；`boundary` 传 null 表示清除。
        有变化时 `version` 加一并写审计日志（`action` = `update`，前后值只含变了的字段）；什么都没变则原样返回，不写日志。
      security:
        - platformToken: []
      parameters:
        - $ref: "#/components/parameters/Id"
        - $ref: "#/components/parameters/IfMatch"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CityPatch"
      responses:
        "200":
          description: 修改后的城市
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/City"
        "404":
          $ref: "#/components/responses/ResourceNotFound"
        "409":
          description: |
            - `VERSION_CONFLICT`：`If-Match` 里的版本号不是最新的，`details.current_version` 是当前版本
            - `FIELD_LOCKED`：改了创建后不能改的字段，`details.fields` 列出是哪些
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "428":
          $ref: "#/components/responses/VersionRequired"
...
  /platform/v1/master/places:
    get:
      tags: [主数据]
      operationId: listPlaces
      summary: 地点列表
      description: |
        需要操作 `master_data.read`（每个平台角色都有）。默认按创建时间从早到晚（创建时间相同的按编号，顺序稳定）；`sort=code` 按编码。
        `total` 是符合筛选条件的总数；`next_cursor` 不为 null 表示还有下一页。`total` 和这一页的内容读自同一时刻的数据。
        不带 `status` 时返回全部（启用的和停用的）。
      security:
        - platformToken: []
      parameters:
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/MasterStatusFilter"
        - $ref: "#/components/parameters/MasterCodeFilter"
        - $ref: "#/components/parameters/MasterKeyword"
        - $ref: "#/components/parameters/MasterSort"
        - $ref: "#/components/parameters/UpdatedSince"
        - $ref: "#/components/parameters/PlaceTypeFilter"
        - $ref: "#/components/parameters/CityIdFilter"
        - $ref: "#/components/parameters/ParentIdFilter"
        - $ref: "#/components/parameters/CountryCodeFilter"
      responses:
        "200":
          description: 一页地点
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [items, next_cursor, total]
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/Place"
                  next_cursor:
                    type: [string, "null"]
                  total:
                    type: integer
                    minimum: 0
                    description: 符合筛选条件的总数（和翻到第几页无关）
                  city_suggestions:
                    type: array
                    description: |
                      只在按 `city_id=none` 筛选时出现：给这一页的每个地点（顺序和 `items` 相同）建议所属城市。
                      候选是和地点同一个国家、启用中的城市，最多 3 个，**不是按距离排的**：
                      数据源说的「机场所属的城市」名称对得上的（150 公里以内）排第一；其余是中心坐标在 80 公里以内的，
                      按「人口 ÷ 距离的平方」从大到小——大城市优先（羽田、成田建议东京，仁川机场建议首尔）。
                      只是建议，请对照 `nearby_cities` 确认。
                    items:
                      $ref: "#/components/schemas/PlaceCitySuggestion"
    post:
      tags: [主数据]
      operationId: createPlace
      summary: 新增地点
      description: |
        需要操作 `master_data.manage`（主数据运营、超级管理员）。
        机场、车站、地标，以及挂在机场下的航站楼、挂在车站下的出口。
        - 编码：机场是 IATA 三字码（`HND`）；车站 `STN-国家码-序号`；地标 `POI-序号`；航站楼 / 出口是上级编码加后缀（`HND-T3`）。全平台唯一。
        - 机场、车站、地标必须给 `city_id`，城市必须是启用的；国家取自城市。
        - 航站楼、出口必须给 `parent_id`（机场 / 车站，必须是启用的）；城市和国家跟随上级，不用给 `city_id`。
        - 车站、地标必须给 `category`；只有地标有 `address`；只有机场和航站楼有 `flight_scope`。
        新增的地点直接是启用的。机场一般不用手工新增，用命令 `pnpm masterdata:import-airports` 从 OurAirports 导入（导入的是停用状态、没有城市）。
        写审计日志（`resource` = `place`，`action` = `create`，`after` 是完整内容）。
      security:
        - platformToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlaceCreate"
      responses:
        "201":
          description: 新增的地点
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Place"
        "409":
          description: |
            - `CODE_TAKEN`：这个编码已被使用
            - `MASTER_DATA_NOT_READY`：所属城市或上级地点已停用（`details.reason` 是 `CITY_DISABLED` / `PARENT_DISABLED`）
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /platform/v1/master/places/{id}:
    get:
      tags: [主数据]
      operationId: getPlace
      summary: 查看一个地点
      description: 需要操作 `master_data.read`。
      security:
        - platformToken: []
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: 地点
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Place"
        "404":
          $ref: "#/components/responses/ResourceNotFound"
    patch:
      tags: [主数据]
      operationId: updatePlace
      summary: 修改地点
      description: |
        需要操作 `master_data.manage`。只改请求体里带了的字段。
        必须用请求头 `If-Match` 带上当前的 `version`：和库里的不一致说明别人先改过了，返回 409 `VERSION_CONFLICT`，不会互相覆盖。
        `type`、`code`、`parent_id` 创建后不能改。`name`、`location` 带了就是整体替换。
        - 换城市：新城市必须和地点在同一个国家；地点是启用的话新城市也必须是启用的。机场 / 车站换城市时，它下面的航站楼 / 出口跟着换。
        - 航站楼、出口不能单独换城市。
        - 导入的机场被改了英文名或坐标后，`source.overridden` 变为 true，以后再导入不会覆盖。
        有变化时 `version` 加一并写审计日志（`action` = `update`，前后值只含变了的字段）；什么都没变则原样返回，不写日志。
      security:
        - platformToken: []
      parameters:
        - $ref: "#/components/parameters/Id"
        - $ref: "#/components/parameters/IfMatch"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlacePatch"
      responses:
        "200":
          description: 修改后的地点
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Place"
        "404":
          $ref: "#/components/responses/ResourceNotFound"
        "409":
          description: |
            - `VERSION_CONFLICT`：`If-Match` 里的版本号不是最新的，`details.current_version` 是当前版本
            - `FIELD_LOCKED`：改了创建后不能改的字段，`details.fields` 列出是哪些
            - `MASTER_DATA_NOT_READY`：地点是启用的，而新城市已停用（`details.reason` 是 `CITY_DISABLED`）
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "428":
          $ref: "#/components/responses/VersionRequired"
...
  /platform/v1/master/vehicle-groups:
    get:
      tags: [主数据]
      operationId: listVehicleGroups
      summary: 车型组列表
      description: |
        需要操作 `master_data.read`（每个平台角色都有）。默认按创建时间从早到晚（创建时间相同的按编号，顺序稳定）；`sort=code` 按编码。
        `total` 是符合筛选条件的总数；`next_cursor` 不为 null 表示还有下一页。`total` 和这一页的内容读自同一时刻的数据。
        不带 `status` 时返回全部（启用的和停用的）。
      security:
        - platformToken: []
      parameters:
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/MasterStatusFilter"
        - $ref: "#/components/parameters/MasterCodeFilter"
        - $ref: "#/components/parameters/MasterKeyword"
        - $ref: "#/components/parameters/MasterSort"
        - $ref: "#/components/parameters/UpdatedSince"
        - $ref: "#/components/parameters/VehicleGradeFilter"
      responses:
        "200":
          description: 一页车型组
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [items, next_cursor, total]
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/VehicleGroup"
                  next_cursor:
                    type: [string, "null"]
                  total:
                    type: integer
                    minimum: 0
                    description: 符合筛选条件的总数（和翻到第几页无关）
    post:
      tags: [主数据]
      operationId: createVehicleGroup
      summary: 新增车型组
      description: |
        需要操作 `master_data.manage`（主数据运营、超级管理员）。
        编码格式 `VG-等级-座位数`（如 `VG-BIZ-7`），末尾的数字必须等于 `seats`。
        等级的缩写是保留的：`ECO` 经济、`CMF` 舒适、`BIZ` 商务、`LUX` 豪华——编码的等级段以其中一个开头时，`grade` 必须就是它
        （`VG-ECO-4` 不能配 `luxury`）；缩写后面可以再加后缀区分变体（`VG-BIZEV-7`）。
        `combos` 是可选的「人数 / 行李数」组合，至少一个，人数不能超过座位数，不能重复。新增的车型组直接是启用的。
        写审计日志（`resource` = `vehicle_group`，`action` = `create`，`after` 是完整内容）。
      security:
        - platformToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VehicleGroupCreate"
      responses:
        "201":
          description: 新增的车型组
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VehicleGroup"
        "409":
          description: |
            - `CODE_TAKEN`：这个编码已被使用
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    platformToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 平台员工登录后拿到的访问令牌
    tenantToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 租户用户登录后拿到的访问令牌；租户编号只从这里取
  parameters:
...
    Limit:
      name: limit
      in: query
      description: 每页条数；只接受十进制整数的写法
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 50
    Cursor:
      name: cursor
      in: query
      description: 上一页返回的 `next_cursor`
      schema:
        type: string
...
    MasterStatusFilter:
      name: status
      in: query
      description: 按状态筛选；`all` 表示不筛选
      schema:
        type: string
        enum: [active, disabled, all]
    MasterCodeFilter:
      name: code
      in: query
      description: 按编码精确查找
      schema:
        type: string
        maxLength: 50
...